Upgrade dependencies and clarify comments
Updated the NEAR SDK to version 4.1.1 to leverage the latest
improvements and fixes. Additionally, incorporated a direct git source
for parity-secp256k1 to ensure compatibility with recent updates. In the
contract code, enhanced the clarity of the developer comments to better
distinguish between view and change methods. Also, bumped the Rust
toolchain version to 1.75.0 to align with current stable releases.

Resolves issue with outdated dependencies.